requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langserve
langsmith
python_dotenv
streamlit
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
wikipedia
streamlit-option-menu
streamlit-extras
PyPDF2